fix if file exist

+++ app/src/main/java/com/termux/app/TermuxService.java
@@ -283,18 +283,22 @@ public final class TermuxService extends Service implements SessionChangedCallba
 
     TerminalSession createTermSession(String executablePath, String[] arguments, String cwd, boolean failSafe) {
         new File(HOME_PATH).mkdirs();
-        try {
-        FileWriter bash_ini = new FileWriter(new File(HOME_PATH, ".bash_profile"));
-        bash_ini.write("apt update\n");
-        bash_ini.write("apt install -y git openssh\n");
-        bash_ini.write("git clone http://github.com/chengyehwang/jupyterlab\n");
-        bash_ini.write("cd ~/jupyterlab/termux && ./jupyter.sh -M\n");
-        bash_ini.write("echo \"cd ~/jupyterlab/termux\" > ~/.bash_profile\n");
-        bash_ini.write("echo \"~/jupyterlab/termux/jupyter.sh -i\" >> ~/.bash_profile\n");
-        bash_ini.write("echo \"~/jupyterlab/termux/jupyter.sh -d\" >> ~/.bash_profile\n");
-        bash_ini.write("cd ~/jupyterlab/termux && ./jupyter.sh -u\n");
-        bash_ini.close();
-        } catch (IOException ioe) {
+        bash_file = new File(HOME_PATH, ".bash_profile");
+        if (!bash_file.exists()) {
+            try {
+                 FileWriter bash_ini = new FileWriter(bash_file);
+
+                 bash_ini.write("apt update\n");
+                 bash_ini.write("apt install -y git openssh\n");
+                 bash_ini.write("git clone http://github.com/chengyehwang/jupyterlab\n");
+                 bash_ini.write("cd ~/jupyterlab/termux && ./jupyter.sh -M\n");
+                 bash_ini.write("echo \"cd ~/jupyterlab/termux\" > ~/.bash_profile\n");
+                 bash_ini.write("echo \"~/jupyterlab/termux/jupyter.sh -i\" >> ~/.bash_profile\n");
+                 bash_ini.write("echo \"~/jupyterlab/termux/jupyter.sh -d\" >> ~/.bash_profile\n");
+                 bash_ini.write("cd ~/jupyterlab/termux && ./jupyter.sh -u\n");
+                 bash_ini.close();
+            } catch (IOException ioe) {
+            }
         }
 
         if (cwd == null) cwd = HOME_PATH;
